move SA key tests to a separate file
