+++ suites/apis/googleDataAccessTest.js
@@ -287,69 +287,3 @@ xScenario('test google data access via usersync: usersync, Google link, generate
     'Cleanup of Google service account for User 2 FAILED.'
   ).to.be.empty;
 });
-
-Scenario('Get temporary Service Account creds, get object in bucket, delete creds @reqGoogle', async (fence, users, google, files) => {
-  const googleProject = fence.props.googleProjectA;
-
-  // Setup
-  await fence.complete.forceLinkGoogleAcct(users.user0, googleProject.owner);
-
-  // Register account
-  const registerRes = await fence.do.registerGoogleServiceAccount(
-    users.user0,
-    googleProject,
-    ['QA']
-  );
-  fence.ask.responsesEqual(registerRes, fence.props.resRegisterServiceAccountSuccess);
-
-  // Get creds to access data
-  const tempCreds0Res = await fence.complete.createTempGoogleCreds(users.user0.accessTokenHeader);
-  const creds0Key = tempCreds0Res.body.private_key_id;
-  const pathToCreds0KeyFile = creds0Key + '.json';
-  await files.createTmpFile(pathToCreds0KeyFile, JSON.stringify(tempCreds0Res.body));
-  console.log(`Google creds file ${pathToCreds0KeyFile} saved!`);
-
-  // Access data
-  user0AccessQARes = await google.getFileFromBucket(
-    fence.props.googleBucketInfo.QA.googleProjectId,
-    pathToCreds0KeyFile,
-    fence.props.googleBucketInfo.QA.bucketId,
-    fence.props.googleBucketInfo.QA.fileName
-  );
-
-  // Delete the key
-  await fence.complete.deleteTempGoogleCreds(
-    creds0Key, users.user0.accessTokenHeader);
-
-  // Try to access data
-  user0AccessQAResExpired = await google.getFileFromBucket(
-    fence.props.googleBucketInfo.QA.googleProjectId,
-    pathToCreds0KeyFile,
-    fence.props.googleBucketInfo.QA.bucketId,
-    fence.props.googleBucketInfo.QA.fileName
-  );
-
-  // Clean up
-  console.log('cleaning up');
-
-  await fence.complete.deleteTempGoogleCreds(
-    creds0Key, users.user0.accessTokenHeader);
-  files.deleteFile(pathToCreds0KeyFile);
-
-  await fence.do.deleteGoogleServiceAccount(
-    users.user0,
-    googleProject.serviceAccountEmail,
-  );
-
-  await fence.do.unlinkGoogleAcct(
-    users.user0,
-  );
-
-  // Asserts
-  chai.expect(user0AccessQARes,
-    'User should have bucket access'
-  ).to.have.property('id');
-  chai.expect(user0AccessQAResExpired,
-    'User should NOT have bucket access after key deletion'
-  ).to.have.property('statusCode', 403);
-});
